+++ src/containers/MapInteractions.js
@@ -71,8 +71,8 @@ export const mapInteractions = compose(
           return eventElement ? eventElement.style.bottom.replace("px", "") : 0;
         })
         .distinctUntilChanged()
-        .subscribe(DraggableContainerHeight =>
-          setMapHeight(Number(DraggableContainerHeight))
+        .subscribe(draggableContainerHeight =>
+          setMapHeight(Number(draggableContainerHeight))
         );
     }
   }),
@@ -86,7 +86,8 @@ export const markerInteractions = compose(
   withRouter,
   connect(selectMarker, { setMapCenter }),
   withHandlers({
-    onClick: ({ id, coords, history, setMapCenter }) => eventId => {
+    onClick: ({ id, coords, history, setMapCenter }) => () => {
+      console.log("CLICKLICK");
       history.push(`/events/${id}`);
       setMapCenter(coords);
     }
